==> notebooks/01_lfr_extraction.ipynb <==
# Extracting and cleaning 2025-2026 London LFR deployment data

- extract data from pdf
- inspect for issues
- convert street adresses to geocode coordinates 

%pip install tabula-py geopy pandas geopandas

%pip install JPype1

from pathlib import Path
import pandas as pd

#obtain clean file path
lfr_file_path = Path(".../data/raw/live-facial-recognition---deployment-record-2025-to-date.pdf")


%pip install pymupdf

# use pymupdf to extract data from table based on visual lines 

import fitz
import pandas as pd

doc = fitz.open(lfr_file_path)
all_rows = []

for i, page in enumerate(doc):
    tabs = page.find_tables()
    
    if tabs.tables:
        raw_data = tabs[0].extract()
        
        if i == 0:
            all_rows.extend(raw_data[2:])
        else:
            all_rows.extend(raw_data[:])
    else:
        print(f"Page {i} is empty or has no table; skipping...")

df_here = pd.DataFrame(all_rows)

doc.close()

print(f"Total rows extracted: {len(df_here)}")

# ok so now I need to add headers + delete first row + delete blank rows 

headers = [
'Deployment Location',
'Date',
'Duration',
'LFR Use Case',
'Watchlist Size',
'Min Threshold Setting',
'Total Alerts',
'True Alerts Confirmed',
'True Alerts Unconfirmed',
'False Alerts Confirmed',
'False Alerts Unconfirmed',
'False Alert Rate',
'Outcome - Arrest',
'Outcome - Other',
'No action',
'Faces seen (estimate)']

df_here.columns = headers

#df_here.drop(index=df_here.index[0], axis=0, inplace=True)

df_here = df_here.dropna(how='any')



df_here.reset_index(drop=True, inplace=True)
df_here['Deployment Location'] = df_here['Deployment Location'].str.replace("\n", " ")
print(df_here.head())

%pip install geopy

#now to geocode 

geo_lfr_df = df_here.copy()

geo_lfr_df["Deployment Location"] = (
    geo_lfr_df["Deployment Location"]
    .str.replace("’", "'", regex=False)
    .str.replace("–", "-", regex=False)
    .str.replace(r"\bSt\b", "Street", regex=True)
    .str.replace(r"\bRd\b", "Road", regex=True)
    .str.replace(r"\bSq\b", "Square", regex=True)
    .str.replace(r"\bStn\b", "Station", regex=True)
    .str.replace("B'Way", "Broadway", regex=False)
    .str.replace("Junc", "Junction", regex=False)
    .str.strip()
)


location_names = geo_lfr_df['Deployment Location'].to_list()

from geopy.geocoders import Nominatim
import time

geolocator = Nominatim(user_agent="lfr_deployment")

latitudes = []
longitudes = []
errors = []

for i, address in enumerate(geo_lfr_df["Deployment Location"]):
    try:
        query = f"{address}, London, UK"
        location = geolocator.geocode(query)
    except Exception as e:
        print(f"Error at row {i}: {e}")
        errors.append(i)
        location = None

    if location:
        latitudes.append(location.latitude)
        longitudes.append(location.longitude)
    else:
        latitudes.append(None)
        longitudes.append(None)
        print(f"Row {i}: '{address}' could not be geocoded")

    time.sleep(1)

geo_lfr_df["latitude"] = latitudes
geo_lfr_df["longitude"] = longitudes

print("Completed")


geo_lfr_df["Deployment Location"] = (
    geo_lfr_df["Deployment Location"]
    .str.replace("Junctiontion", "Junction", regex=False)
    .str.replace("B'way", "Broadway", regex=False)
)

# manually assing the missing rows 
# https://www.gps-coordinates.net/

geo_lfr_df.loc[26, ["latitude", "longitude"]] = [51.506255, -0.220575]
geo_lfr_df.loc[39, ["latitude", "longitude"]] = [51.51498, -0.300407]
geo_lfr_df.loc[87, ["latitude", "longitude"]] = [51.427739, -0.16829]
geo_lfr_df.loc[94, ["latitude", "longitude"]] = [51.51498, -0.300407]
geo_lfr_df.loc[98, ["latitude", "longitude"]] = [51.511903, -0.014372]
geo_lfr_df.loc[102, ["latitude", "longitude"]] = [51.531707, -0.124766]
geo_lfr_df.loc[103, ["latitude", "longitude"]] = [51.506295, -0.231435]
geo_lfr_df.loc[119, ["latitude", "longitude"]] = [51.509759, -0.13355]
geo_lfr_df.loc[130, ["latitude", "longitude"]] = [51.544337, -0.00632]
geo_lfr_df.loc[171, ["latitude", "longitude"]] = [51.543793, 0.049819]
geo_lfr_df.loc[198, ["latitude", "longitude"]] = [51.509759, -0.13355]
geo_lfr_df.loc[207, ["latitude", "longitude"]] = [51.507459, -0.222272]
geo_lfr_df.loc[227, ["latitude", "longitude"]] = [51.545284, -0.074968]

#need to complete the rest

# number of locations were manually assigned coordinates
#“Locations that failed automated geocoding were reviewed and standardised manually before being reprocessed.”
#https://www.gps-coordinates.net/



# FInal check 

missing_values = geo_lfr_df.isnull().sum()

print(missing_values)

# Data is correct and complete



#save document

data_dir = Path("../data/processed")
data_dir.mkdir(exist_ok=True)

output_path = data_dir / "lfr_clean_geo.csv"
geo_lfr_df.to_csv(output_path, index=False)


# Convert to geopackadge

from pathlib import Path
import pandas as pd

lfr_df = pd.read_csv(Path("../data/processed/lfr_clean_geo.csv"))


#convert to geopackadge

import geopandas as gpd
from shapely.geometry import Point

lfr_gdf = gpd.GeoDataFrame(
    lfr_df,
    geometry=[Point(xy) for xy in zip(lfr_df.longitude, lfr_df.latitude)],
    crs="EPSG:4326"
)




lfr_output = Path("../data/processed/lfr_deployments.gpkg")

lfr_gdf.to_file(
    lfr_output,
    layer="lfr_deployments",
    driver="GPKG"
)
