==> notebooks/02_stop_search_extraction.ipynb <==
# Extracting stop and search data and checking for LSOA codes

- check 2025 stop and search data
- Check for LSOA codes and clean 
- save to processed data 

from pathlib import Path
import pandas as pd




data_dir = Path("../data/raw")
csv_files = sorted(data_dir.glob("2025-*-metropolitan-stop-and-search.csv"))
csv_files

dataframes = []

for file in csv_files:
    df = pd.read_csv(file)
    dataframes.append(df)

stop_and_search_2025 = pd.concat(dataframes, ignore_index=True)



stop_and_search_2025.head()

#OK no LSOA but there is coordinates 
#Data appears clean and ready to use. 

#save document

data_dir = Path("../data/processed")
data_dir.mkdir(exist_ok=True)

output_path = data_dir / "stop_search_aggregated.csv"
stop_and_search_2025.to_csv(output_path, index=False)
