==> notebooks/03_LSOA_aggregation.ipynb <==
# LSOA aggreagation 
- extract the LSOA boundaries shapefile 
- aggregate with 2025 stop and search data to create choloropleth 


from pathlib import Path
import geopandas as gpd


zip_path = Path("../data/raw/statistical-gis-boundaries-london.zip")

shp_inside_zip = "statistical-gis-boundaries-london/ESRI/LSOA_2011_London_gen_MHW.shp"



lsoa = gpd.read_file(f"zip://{zip_path}!{shp_inside_zip}")

# Check
print(lsoa.head())
print(lsoa.crs)


#OK need to convert the CRS to work with lat long


lsoa = lsoa.to_crs("EPSG:4326")

#import stop and search data 

import pandas as pd

processed_dir = Path("../data/processed")
stop_search_file = processed_dir / "stop_search_aggregated.csv"

stop_and_search_2025_data = pd.read_csv(stop_search_file)

print(stop_and_search_2025_data.head())


#Convert dataframe to geodatafram

import geopandas as gpd
from shapely.geometry import Point

gdf = gpd.GeoDataFrame(
    stop_and_search_2025_data,
    geometry=[Point(xy) for xy in zip(stop_and_search_2025_data.Longitude, stop_and_search_2025_data.Latitude)],
    crs="EPSG:4326"
)


# Add LSOA data to geodataframe

joined = gpd.sjoin(
    gdf,
    lsoa,
    how="left",
    predicate="within"
)



#Check

print(joined[["LSOA11CD"]].head())

#Create counts for each area 

lsoa_counts = (
    joined
    .groupby("LSOA11CD")
    .size()
    .reset_index(name="stop_search_count")
)


# Add the counts to the polygons 

lsoa_map = lsoa.merge(
    lsoa_counts,
    on="LSOA11CD",
    how="left"
).fillna(0)


# Export the lsoa as geopackadge 

from pathlib import Path

output_dir = Path("../data/processed")

lsoa_output = output_dir / "lsoa_stop_search.gpkg"

lsoa_map.to_file(
    lsoa_output,
    layer="lsoa_stop_search",
    driver="GPKG"
)
