==> notebooks/04_hist_s&s_extraction.ipynb <==
# Extract historical stop and search data to display change over time 

- import Period jan to nov 2023 and concat idnivdual files 
- LFR deployment locations reflect deployments occurring at any point during 2025, while stop and search activity is aggregated for January–November 2025 to ensure comparability with earlier periods.


from pathlib import Path
import pandas as pd




data_dir = Path("../data/raw/stop_search_jan_nov_2023")
csv_files = sorted(data_dir.glob("2023-*-metropolitan-stop-and-search.csv"))
csv_files

dataframes = []

for file in csv_files:
    df = pd.read_csv(file)
    dataframes.append(df)

stop_and_search_2023 = pd.concat(dataframes, ignore_index=True)



#save document

data_dir = Path("../data/processed")
data_dir.mkdir(exist_ok=True)

output_path = data_dir / "2023_stop_search_aggregated.csv"
stop_and_search_2023.to_csv(output_path, index=False)
