==> notebooks/06_deprivation_extract_join.ipynb <==
# Deprivation data mapping
- extract london data deprivation by LSOA
- join to LSOA polygons 

from pathlib import Path 
import pandas as pd

dep_path = Path('../data/raw/File_1_-_IMD2019_Index_of_Multiple_Deprivation.xlsx')

deprivation_df = pd.read_excel(dep_path, sheet_name='IMD2019')

# Only extract the LSOAs under investigation 
# Use the LSOA list previously extracted to match against 
import geopandas as gpd
 
lso_path = Path('../data/processed/lsoa_change_2023_2025.gpkg')


lsoa_gdf = gpd.read_file(lso_path)

#Filter so that only london is included 

rel_lsoas = lsoa_gdf['LSOA11CD'].unique()

dep_london = deprivation_df[deprivation_df['LSOA code (2011)'].isin(rel_lsoas)]

dep_london.shape

#rename to join
dep_london = dep_london.rename(columns={'LSOA code (2011)': 'LSOA11CD'})

# merge on LSOAs

merged_dep = lsoa_gdf.merge(dep_london[['LSOA11CD', 'Index of Multiple Deprivation (IMD) Decile']], how='left', on='LSOA11CD')

# check 

merged_dep["Index of Multiple Deprivation (IMD) Decile"].isna().sum()

merged_dep.to_file('../data/processed/imd_dep_london.gpkg')
